requirements.txt
#main_openai
PyPDF2~=3.0.1
langchain~=0.0.231
openai~=0.27.8
tiktoken~=0.4.0
faiss-cpu~=1.7.4
python-dotenv~=1.0.0
#pinecone-client~=2.2.1
#sympy~=


#main_gpt4all
gpt4all~=1.0.8
sentence_transformers

#lettura pagine html
requests~=2.31.0
BeautifulSoup4~=4.12.2
lxml~=4.9.3

#main_bard
bardapi~=0.1.32

#main_huggingface
#langchain~=0.0.300
huggingface_hub=0.16.4
sentence_transformers=2.2.2
